Avoid redundant build run
